Resolve app import cycles
