+++ packages/app/ui/components/Activity/ActivityListItem.tsx
@@ -9,7 +9,8 @@ import { XStack, YStack, styled } from 'tamagui';
 
 import { useCalm } from '../../contexts/appDataContext';
 import { getGroupTitle, useChannelTitle, useGroupTitle } from '../../utils';
-import { ChannelAvatar, ContactAvatar, GroupAvatar } from '../Avatar';
+import { ChannelAvatar, ContactAvatar } from '../Avatar';
+import { GroupAvatar } from '../GroupAvatar';
 import { UnreadDot } from '../UnreadDot';
 import { ActivitySourceContent } from './ActivitySourceContent';
 import { SummaryMessage } from './ActivitySummaryMessage';
