+++ packages/app/ui/components/Avatar.tsx
@@ -18,9 +18,7 @@ import { useCalm, useContact } from '../contexts/appDataContext';
 import * as utils from '../utils';
 import { getChannelTypeIcon } from '../utils';
 import { getContrastingColor, useSigilColors } from '../utils/colorUtils';
-import { FacePile } from './FacePile';
-
-const AvatarFrame = styled(View, {
+export const AvatarFrame = styled(View, {
   width: '$4xl',
   height: '$4xl',
   borderRadius: '$s',
@@ -115,89 +113,6 @@ export interface GroupImageShim {
   title?: string;
   iconImage?: string;
   iconImageColor?: string;
-}
-const SMALL_AVATAR_SIZES = ['$xl', '$2xl', '$3xl', '$3.5xl'] as const;
-
-export const GroupAvatar = React.memo(function GroupAvatarComponent({
-  model,
-  memberCount,
-  membersLayout = 'default',
-  ...props
-}: {
-  model: db.Group | GroupImageShim;
-  memberCount?: number;
-  membersLayout?: 'default' | 'compact';
-} & AvatarProps) {
-  const { disableNicknames } = useCalm();
-  const fallbackTitle = useMemo(() => {
-    return isGroupImageShim(model)
-      ? model.title
-      : utils.getGroupTitle(model, disableNicknames);
-  }, [disableNicknames, model]);
-
-  const memberContactIds = useMemo(() => {
-    if (isGroupImageShim(model)) {
-      return [];
-    }
-    return model.members?.map((m) => m.contactId) ?? [];
-  }, [model]);
-
-  const isSmallSize = SMALL_AVATAR_SIZES.includes(
-    (props.size ?? '$4xl') as (typeof SMALL_AVATAR_SIZES)[number]
-  );
-  const facePileGridDensity =
-    membersLayout === 'compact' ? 'compact' : 'default';
-
-  const textFallback = (
-    <TextAvatar
-      text={fallbackTitle ?? 'G'}
-      backgroundColor={model.iconImageColor ?? undefined}
-      {...props}
-    />
-  );
-
-  const fallback =
-    memberContactIds.length > 0 ? (
-      isSmallSize ? (
-        <FacePile
-          contactIds={memberContactIds}
-          maxVisible={2}
-          totalCount={memberCount}
-        />
-      ) : (
-        <AvatarFrame
-          {...props}
-          alignItems="center"
-          justifyContent="center"
-          backgroundColor="$secondaryBackground"
-        >
-          <FacePile
-            contactIds={memberContactIds}
-            maxVisible={4}
-            totalCount={memberCount}
-            grid
-            gridDensity={facePileGridDensity}
-          />
-        </AvatarFrame>
-      )
-    ) : (
-      textFallback
-    );
-
-  return (
-    <ImageAvatar
-      imageUrl={model.iconImage ?? undefined}
-      fallback={fallback}
-      isGroupIcon={true}
-      {...props}
-    />
-  );
-});
-
-function isGroupImageShim(
-  group: db.Group | GroupImageShim
-): group is GroupImageShim {
-  return !('description' in group);
 }
 
 export const ChannelAvatar = React.memo(function ChannelAvatarComponent({
