+++ packages/app/ui/components/Channel/scrollerTypes.ts
@@ -1,0 +1,4 @@
+export type ScrollAnchor = {
+  type: 'unread' | 'selected';
+  postId: string;
+};
