+++ packages/app/ui/components/Embed/EmbedContent.tsx
@@ -5,7 +5,7 @@ import { memo, useCallback, useMemo } from 'react';
 import { Linking, Platform } from 'react-native';
 
 import { useCalm } from '../../contexts/appDataContext';
-import { AudioEmbed } from '../Embed';
+import AudioEmbed from './AudioEmbed';
 import { Embed } from './Embed';
 import { EmbedWebView } from './EmbedWebView';
 import { getProviderConfig } from './providers';
