+++ packages/app/ui/components/NotebookPost/NotebookPost.tsx
@@ -26,7 +26,7 @@ import {
 import { useBlockedAuthor } from '../../../hooks/useBlockedAuthor';
 import { useCurrentUserId } from '../../contexts/appDataContext';
 import { useChannelContext } from '../../contexts/channel';
-import type { MinimalRenderItemProps } from '../../contexts/componentsKits';
+import type { MinimalRenderItemProps } from '../../contexts/componentsKits/componentsKits';
 import { useCanWrite } from '../../utils/channelUtils';
 import { DetailViewAuthorRow } from '../AuthorRow';
 import { ChatMessageActions } from '../ChatMessage/ChatMessageActions/Component';
