+++ packages/app/ui/index.tsx
@@ -4,6 +4,8 @@ export * from './components/AddContactsView';
 export * from './components/AppSetting';
 export * from './components/ArvosDiscussing';
 export * from './components/Avatar';
+export { GroupAvatar } from './components/GroupAvatar';
+export type { GroupImageShim } from './components/Avatar';
 export * from './components/BigInput';
 export * from './components/BlockedContactsWidget';
 export * from './components/Channel';
