+++ packages/app/utils/identifyUser.ts
@@ -1,10 +1,8 @@
 import * as Sentry from '@sentry/react-native';
 
-import { posthog } from './posthog';
-
 /**
- * Identifies a user across both PostHog and Sentry telemetry systems.
- * This ensures user context stays synchronized between both platforms.
+ * Identifies a user in Sentry telemetry. PostHog identification should be
+ * done separately via the posthog client to avoid circular dependencies.
  *
  * @param userId - The user identifier (typically ship ID like ~sampel-palnet)
  * @param properties - Additional user properties to attach
@@ -13,10 +11,7 @@ export function identifyUser(
   userId: string,
   properties?: { isHosted?: boolean; [key: string]: any }
 ) {
-  // Update PostHog user identification
-  posthog?.identify(userId, properties);
-
-  // Update Sentry user context to match
+  // Update Sentry user context
   Sentry.setUser({
     id: userId,
     ...properties,
